fixes the routing and authentications
