+++ helpComponents/LandingNavbar.tsx
@@ -24,11 +24,11 @@ export default function LandingNavbar(){
               <a href="#features" className="text-foreground/80 hover:text-primary transition-smooth">Features</a>
               <a href="#benefits" className="text-foreground/80 hover:text-primary transition-smooth">Benefits</a>
               <a href="#testimonials" className="text-foreground/80 hover:text-primary transition-smooth">Reviews</a>
-              <Link href="/signup">
-                <Button variant="outline" className="mr-4">Log In</Button>
-              </Link>
               <Link href="/signin">
-                <Button className="gradient-primary text-white border-0 shadow-primary hover:shadow-glow transition-smooth">Get Started</Button>
+                <Button variant="outline" className="mr-4 cursor-pointer ">Log In</Button>
+              </Link>
+              <Link href="/signup">
+                <Button className="cursor-pointer gradient-primary text-white border-0 shadow-primary hover:shadow-glow transition-smooth">Get Started</Button>
               </Link>
             </div>
           </div>
