--- helpComponents/LoginForm.tsx
@@ -1,183 +1,0 @@
-import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Separator } from "@/components/ui/separator";
-import { Eye, EyeOff, Mail, Lock, Chrome } from "lucide-react";
-import { toast } from "sonner";
-
-const LoginForm = () => {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [showPassword, setShowPassword] = useState(false);
-  const [rememberMe, setRememberMe] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setIsLoading(true);
-    
-    // Simulate login process
-    await new Promise(resolve => setTimeout(resolve, 1500));
-    
-    toast.success("Welcome back! Login successful.");
-    setIsLoading(false);
-  };
-
-  const handleSocialLogin = (provider: string) => {
-    if (provider === "Google") {
-      window.location.href = "https://accounts.google.com/oauth/authorize?client_id=your-google-client-id&redirect_uri=your-redirect-uri&scope=openid%20email%20profile&response_type=code";
-    } else if (provider === "LinkedIn") {
-      window.location.href = "https://www.linkedin.com/oauth/v2/authorization?client_id=your-linkedin-client-id&redirect_uri=your-redirect-uri&scope=r_liteprofile%20r_emailaddress&response_type=code";
-    }
-  };
-
-  return (
-    <div className="w-full max-w-md mx-auto">
-      <div className="card-gradient rounded-2xl p-8 animate-slide-up">
-        <div className="text-center mb-8">
-          <h1 className="text-3xl font-bold text-foreground mb-2">
-            Welcome Back
-          </h1>
-          <p className="text-muted-foreground">
-            Sign in to your contractor portal
-          </p>
-        </div>
-
-        <form onSubmit={handleSubmit} className="space-y-6">
-          <div className="space-y-2">
-            <Label htmlFor="email" className="text-sm font-medium">
-              Email Address
-            </Label>
-            <div className="relative">
-              <Mail className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-4 w-4" />
-              <Input
-                id="email"
-                type="email"
-                placeholder="contractor@example.com"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                className="pl-10 h-12 border-border focus:ring-primary focus:border-primary transition-all duration-200"
-                required
-              />
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <Label htmlFor="password" className="text-sm font-medium">
-              Password
-            </Label>
-            <div className="relative">
-              <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-4 w-4" />
-              <Input
-                id="password"
-                type={showPassword ? "text" : "password"}
-                placeholder="Enter your password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                className="pl-10 pr-10 h-12 border-border focus:ring-primary focus:border-primary transition-all duration-200"
-                required
-              />
-              <button
-                type="button"
-                onClick={() => setShowPassword(!showPassword)}
-                className="absolute right-3 top-1/2 transform -translate-y-1/2 text-muted-foreground hover:text-foreground transition-colors"
-              >
-                {showPassword ? <EyeOff className="h-4 w-4" /> : <Eye className="h-4 w-4" />}
-              </button>
-            </div>
-          </div>
-
-          <div className="flex items-center justify-between">
-            <div className="flex items-center space-x-2">
-              <Checkbox
-                id="remember"
-                checked={rememberMe}
-                onCheckedChange={(checked) => setRememberMe(checked as boolean)}
-              />
-              <Label htmlFor="remember" className="text-sm text-muted-foreground">
-                Remember me
-              </Label>
-            </div>
-            <a
-              href="#"
-              className="text-sm text-primary hover:text-primary-glow transition-colors font-medium"
-              onClick={(e) => {
-                e.preventDefault();
-                toast.info("Password reset link sent to your email!");
-              }}
-            >
-              Forgot password?
-            </a>
-          </div>
-
-          <Button
-            type="submit"
-            className="w-full h-12 button-gradient hover:hover-glow text-primary-foreground font-semibold transition-all duration-300 hover-lift"
-            disabled={isLoading}
-          >
-            {isLoading ? (
-              <div className="flex items-center space-x-2">
-                <div className="w-4 h-4 border-2 border-primary-foreground border-t-transparent rounded-full animate-spin" />
-                <span>Signing in...</span>
-              </div>
-            ) : (
-              "Sign In"
-            )}
-          </Button>
-        </form>
-
-        <div className="mt-8">
-          <div className="relative">
-            <div className="absolute inset-0 flex items-center">
-              <Separator className="w-full" />
-            </div>
-            <div className="relative flex justify-center text-xs uppercase">
-              <span className="bg-card px-2 text-muted-foreground">Or continue with</span>
-            </div>
-          </div>
-
-          <div className="mt-6 grid grid-cols-2 gap-3">
-            <Button
-              variant="outline"
-              onClick={() => handleSocialLogin("Google")}
-              className="h-12 border-border hover:bg-muted hover-lift transition-all duration-200"
-            >
-              <Chrome className="h-4 w-4 mr-2" />
-              Google
-            </Button>
-            <Button
-              variant="outline"
-              onClick={() => handleSocialLogin("LinkedIn")}
-              className="h-12 border-border hover:bg-muted hover-lift transition-all duration-200"
-            >
-              <svg className="h-4 w-4 mr-2" fill="currentColor" viewBox="0 0 24 24">
-                <path d="M20.447 20.452h-3.554v-5.569c0-1.328-.027-3.037-1.852-3.037-1.853 0-2.136 1.445-2.136 2.939v5.667H9.351V9h3.414v1.561h.046c.477-.9 1.637-1.85 3.37-1.85 3.601 0 4.267 2.37 4.267 5.455v6.286zM5.337 7.433c-1.144 0-2.063-.926-2.063-2.065 0-1.138.92-2.063 2.063-2.063 1.14 0 2.064.925 2.064 2.063 0 1.139-.925 2.065-2.064 2.065zm1.782 13.019H3.555V9h3.564v11.452zM22.225 0H1.771C.792 0 0 .774 0 1.729v20.542C0 23.227.792 24 1.771 24h20.451C23.2 24 24 23.227 24 22.271V1.729C24 .774 23.2 0 22.222 0h.003z"/>
-              </svg>
-              LinkedIn
-            </Button>
-          </div>
-        </div>
-
-        <div className="mt-8 text-center">
-          <p className="text-sm text-muted-foreground">
-            Don't have an account?{" "}
-            <a
-              href="#"
-              className="text-primary hover:text-primary-glow font-medium transition-colors"
-              onClick={(e) => {
-                e.preventDefault();
-                toast.info("Sign up page coming soon!");
-              }}
-            >
-              Sign up for free
-            </a>
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default LoginForm;
